Move more related items to useResize.
